Revert "chore: remove balance check before executing route"
This reverts commit 39cf7bba4966a1dd8bd7671be00f69df0276b4e3.
